modified: - buidlid now removed - added also linux (even if not supported in sliver) - extension.json
